Renamed route constants file
