+++ project/src/store/action.ts
@@ -1,5 +1,5 @@
 import { createAction } from '@reduxjs/toolkit';
-import { WebRoutes } from '../constants/webRoutes';
+import { WebRoutes } from '../constants/routes';
 
 export const changeGenre = createAction<string>('changeGenre');
 export const redirectToRoute = createAction<WebRoutes | string>('redirectToRoute');
