+++ project/src/store/reducer/reducer.ts
@@ -2,7 +2,7 @@ import { mainReducer } from './main/main-reducer';
 import { filmReducer } from './film/film-reducer';
 import { userReducer } from './user/user-reducer';
 import { combineReducers } from '@reduxjs/toolkit';
-import { Namespace } from '../../constants/webRoutes';
+import { Namespace } from '../../constants/routes';
 
 export const reducer = combineReducers({
   [Namespace.Data]: mainReducer.reducer,
